+++ NEWS.md
@@ -1,5 +1,7 @@
 # bstfun (development version)
 
+* Added `tbl_split()` function to actively split a gtsummary table to more easily show on multiple pages.
+
 * Adding `as_ggplot()` function to convert gt and gtsummary tables to ggplot
 
 * Added `"leej"` theme to `theme_gtsummary_msk()`